refactor: rename adapter module to transform
